refactor(institutions/product): change model and update test (part1)
