+++ src/main/java/it/pagopa/pn/bff/mappers/CxTypeMapper.java
@@ -14,5 +14,5 @@ public interface CxTypeMapper {
     it.pagopa.pn.bff.generated.openapi.msclient.delivery_pa.model.CxTypeAuthFleet convertDeliveryPACXType(CxTypeAuthFleet cxType);
 
     it.pagopa.pn.bff.generated.openapi.msclient.apikey_pa.model.CxTypeAuthFleet convertApiKeysPACXType(CxTypeAuthFleet cxType);
-    it.pagopa.pn.bff.generated.openapi.msclient.external_registries.model.CxTypeAuthFleet convertExternalRegistriesCXType(CxTypeAuthFleet cxType);
+    it.pagopa.pn.bff.generated.openapi.msclient.external_registries_selfcare.model.CxTypeAuthFleet convertExternalRegistriesCXType(CxTypeAuthFleet cxType);
 }
